+++ src/main/java/org/my/quarkus/raft/api/RequestVoteRequest.java
@@ -1,4 +1,4 @@
 package org.my.quarkus.raft.api;
 
-public record RequestVoteRequest(long term, String candidateId, long lastLogIndex, long lastLogTerm) {
+public record RequestVoteRequest(int term, String candidateId, int lastLogIndex, int lastLogTerm) {
 }
