+++ src/main/java/org/my/quarkus/raft/api/RequestVoteResponse.java
@@ -1,4 +1,4 @@
 package org.my.quarkus.raft.api;
 
-public record RequestVoteResponse(long term, boolean voteGranted) {
+public record RequestVoteResponse(int term, boolean voteGranted) {
 }
